+++ src/Notify/Models/Template.cs
@@ -5,5 +5,20 @@
         public string id;
         public string uri;
         public int version;
+
+        public override bool Equals(object other)
+        {
+            if (!(other is Template o))
+            {
+                return false;
+            }
+
+            return id == o.id && uri == o.uri && version == o.version;
+        }
+
+        public override int GetHashCode()
+        {
+            return base.GetHashCode();
+        }
     }
 }